Added new GetBankAccountByIdQuery and refactored GetBankAccountsQuery
