+++ Queries/BankAccounts/BankAccountDetail.cs
@@ -1,0 +1,3 @@
+namespace Queries.BankAccounts;
+
+public record BankAccountDetail(Guid Id, string Name, string AccountNumber);
