+++ Queries/BankAccounts/GetBankAccounts.cs
@@ -1,7 +1,27 @@
-using MediatR;
+using Microsoft.EntityFrameworkCore;
 
-using Queries.BankAccounts.ViewModels;
+using Persistence;
 
-namespace Queries;
+using Queries.BankAccounts;
 
-public record GetBankAccounts() : IRequest<ICollection<BankAccount>>;
+namespace Queries.GetBankAccounts;
+
+public record GetBankAccounts() : IRequest<ICollection<BankAccountDetail>>;
+
+public class GetBankAccountsHandler : IRequestHandler<GetBankAccounts, ICollection<BankAccountDetail>>
+{
+    private readonly IDbContextFactory<HaSpManContext> _contextFactory;
+
+    public GetBankAccountsHandler(IDbContextFactory<HaSpManContext> contextFactory)
+    {
+        _contextFactory = contextFactory;
+    }
+
+    public async Task<ICollection<BankAccountDetail>> Handle(GetBankAccounts request, CancellationToken cancellationToken)
+    {
+        var context = _contextFactory.CreateDbContext();
+        return await context.BankAccounts
+            .Select(b => new BankAccountDetail(b.Id, b.Name, b.AccountNumber))
+            .ToListAsync(cancellationToken);
+    }
+}
