--- Queries/BankAccounts/ViewModels/BankAccountDetail.cs
@@ -1,3 +1,0 @@
-namespace Queries.BankAccounts.ViewModels;
-
-public record BankAccount(Guid Id, string Name, string AccountNumber);
