+++ Queries/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using AutoMapper;
+
+global using MediatR;
